+++ lib/rubocop/cop/rspec/capybara/match_style.rb
@@ -29,30 +29,8 @@ module RuboCop
         #   # good
         #   expect(page).to match_style(display: 'block')
         #
-        class MatchStyle < Base
-          extend AutoCorrector
-
-          MSG = 'Use `%<good>s` instead of `%<bad>s`.'
-          RESTRICT_ON_SEND = %i[assert_style has_style? have_style].freeze
-          PREFERRED_METHOD = {
-            'assert_style' => 'assert_matches_style',
-            'has_style?' => 'matches_style?',
-            'have_style' => 'match_style'
-          }.freeze
-
-          def on_send(node)
-            method_node = node.loc.selector
-            add_offense(method_node) do |corrector|
-              corrector.replace(method_node,
-                                PREFERRED_METHOD[method_node.source])
-            end
-          end
-
-          private
-
-          def message(node)
-            format(MSG, good: PREFERRED_METHOD[node.source], bad: node.source)
-          end
+        # @!parse class MatchStyle < ::RuboCop::Cop::Base; end
+        class MatchStyle < ::RuboCop::Cop::Capybara::MatchStyle
         end
       end
     end