+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Fix a false negative for `RSpec/Pending` when using skipped in metadata is multiline string. ([@ydah])
 - Fix a false positive for `RSpec/NoExpectationExample` when using skipped in metadata is multiline string. ([@ydah])
 - Fix a false positive for `RSpec/ContextMethod` when multi-line context with `#` at the beginning. ([@ydah])
+- Extract Capybara cops to a separate repository. ([@pirj])
 
 ## 2.17.0 (2023-01-13)